Release build 3

+++ soundlib/ModChannel.h
@@ -124,7 +124,7 @@ struct ModChannel
 	uint8 nRestoreResonanceOnNewNote;  // See nRestorePanOnNewNote
 	uint8 nRestoreCutoffOnNewNote;     // ditto
 	uint8 nNote;
-	uint8 noteTickDelay;
+	int16 noteTickDelay;
 	NewNoteAction nNNA;
 	uint8 nLastNote;  // Last note, ignoring note offs and cuts - for MIDI macros
 	uint8 nArpeggioLastNote, lastMidiNoteWithoutArp;  // For plugin arpeggio and NNA handling
